+++ src/components/Contact.jsx
@@ -21,6 +21,7 @@ const Contact = ({ author }) => {
           </div>
         </div>
       </div>
+      
     </>
   );
 };